Makes ValueSetGenerator work for both STU3 and R4.
Improvements to ValueSetGenerator API:  Adds in --for_codes_in and --exclude_codes_in which take structure definition bundles.  Instead of having to explicitly list the codes to include, they can be inferred by reading structure definitions and determining which codes will be necessary to inlined, as well as excluded.

The use case for excluded is for generating codes for a new profile without generating all the codes in the core fhir, as well as saying "all codes in the resources/profiles/extensions, but EXCLUDE the ones in datatypes, since they must be added to the datatypes.proto itself to avoid circular dependencies.

PiperOrigin-RevId: 251947282