Makes json_format.h correctly parse to and print from profiled protos.
All output will be valid FHIR JSON, and parsing to a profile will fail if the resource does not match the profile.

Finally, this adds a validation step (along with a no-validate variant) to parsing json to a non-profiled proto.  This guarantees that unless explicitly requested, ANY proto parsed from JSON is valid FHIR.

PiperOrigin-RevId: 275744604