Update of Code handling in R4 protos.  Now only enums will be generated in codes.proto, and bound codes (along with id and extension) will be generated in line where they are used.
ValueSet and CodeSystem resources are piped into the ProtoGenerator, so that even if a code system was not previously generated for a binding, it can be generated on the fly in the proto itself, meaning any value set that is defined by a resource can be generated.

Note that this change results in no binary incompatible changes in the core resources/datatypes.

PiperOrigin-RevId: 276588230