Implement FHIRPath's children() function.
Also, change the behavior of the evaluation engine when unknown fields are accessed. Specifically, if it is not clear at compile time that the field being accessed does not exist it will compile and will result in a empty collection during evaluation (assuming the field does not, in fact, exist.) Depending on the response to https://chat.fhir.org/#narrow/stream/179266-fhirpath/topic/Path.20selection/near/187600058 I may change the behavior for situations where we know the field will never exist to match the case where we don't know until evaluation time (i.e. always return an empty collection.)

PiperOrigin-RevId: 294335055