Removes FhirProto dependency on Tensorflow.
In addition to removing the org_tensorflow dependency (and the very few places it was used), this eliminates the tf_workspace call in fhir_workspace, and replaces all dependencies declared from it with explicit dependencies declared in fhir_workspace.  The only modification to dependencies is replacing tensorflow logging with glog.

PiperOrigin-RevId: 334463559