Removes always-false branch and marks bug + TODO for minor issue in CopyCoding.
This assumes that the system of a profiled Coding can always be determined by inspecting the profiled code field.  This is not technically true if you have a profiled Coding that does not have a profiled Code field.  FhirProto does not currently support such profiles, but they are theoretically possible according to FHIR.  The removed branch checks for a "system" field on a Code, which never exists - the code field exists on the containing Coding.

PiperOrigin-RevId: 353622812