Modifies ProtoGeneratorMain's handling of FileSplittingBehavior to *ALWAYS* output only a single .zip file containing all resources, codes, and extensions, regardless of how those are subdivided within the zip.
This allows bazel to be totally ignorant of what actual .proto files will be generated, and allows greater flexibility in the generator itself for choosing what files are appropriate.  As a result, the bazel flag "separate_extensions" that duplicated information from PackageInfo is removed, along with some flags to ProtoGeneratorMain - previously the main point of these flags was to make sure bazel knew what files to expect.

This also has the side effect of generating the new src files with 664 permissions bitmask instead of 775.  This is effectively a bug fix - there's no reason these files should have executable bits.

PiperOrigin-RevId: 367738377