Replaces old GeneratedProtoTest regression test with a much more powerful auto regression test.
The original GeneratedProtoTest attempted to replicate the arguments used to generate the protos originally, run them through the ProtoGenerator (but not ProtoGeneratorMain), and compare with a FileDescriptorProto from a runtime dep.  This had no guarantees that the arguments were correct, and did not test extensions, codes, or split resource files.

The new GeneratedProtoTest instead uses a direct dependency on the ephemeral bazel-genfiles zip, and compares that directly with the .proto files on disk.  The bazel dep on a genrule guarantees that the test files are up to date and identical to the output of running the protogeneration (regardless of any changes to flags).  Using the .proto files on disk as goldens guarantees comprehensive coverage.

PiperOrigin-RevId: 368140846