Remove environment marker for backports.zoneinfo when using Bazel.
The `rules_python` `requirement` macro only does name-mangling. To handle
an environment marker for a py_library dependency we will need something more
sophisticated (e.g. a macro that accounts for the interpreter version).

Since Bazel requirements already force the user to have the tools necessary to
build a backports.zoneinfo whl, for now we'll keep this pip-dependency for
Bazel, but have a PEP 508 environment marker for our pypi distribution.

PiperOrigin-RevId: 380644314