Update ValidationOutcome.subject to be optional.
This is because we may need to validate FHIR resources that do not have an ID set. In this case, we can't set ValidationOutcome.subject to something meaningful, but still would like to produce a ValidationOutcome (e.g. from APIs like Validate() [1]).

In batch settings, producing ValidationOutcomes that can't point to a resource may be meaningless--but for other uses of the Validate API this isn't the case so having this as optional will still be useful.

PiperOrigin-RevId: 394257563