Change to ErrorReporter C++ API to make it easier to differentiate failures results vs errors encountered, and allow for reporting both the field that failed, and the exact element (including indices into repeated fields).
This changes the APIs from
{validation error, conversion error, fhirpath error, fhirpath warning}
to
{warning, failure, error, fhirpath warning, fhirpath failure, fhirpath error}

This is a total no-op with one exception:  Previously an invalid Reference was being reported as a ConversionError, which would map to an Error in the new system since failures in conversion mean the output data cannot be generated.  Invalid References are now correctly reported as failures rather than errors.

PiperOrigin-RevId: 430288828