Retrieve Current Value Set Definition Alongside Expansion
TerminologyServiceClient.expand_value_set now returns the current
definition of the value set from the terminology service rather than
appending the expanded codes to the passed value set object. This is
particularly useful if passed a partial value set with only the URL
value present, as it allows the full definition of the value set to be
retrieved.

PiperOrigin-RevId: 430996333