Pipes ErrorReporter throughout JSON parsing in C++.
Maintains all external APIs backwards compatible, but converts internal parse functions from returning "Status" (where OK means "everything parsed and the data was valid", and not ok means "Either data was invalid or we encountered an internal error") to returning "StatusOr<ParseResult>", where ParseResult indicates if the parse was successful (lossless) or unsuccessful (lossy/incomplete), and a status indicates an internal code error unrelated to user data.

In this model, invalid data does not necessarily cause the parse to fail.  For instance, when parsing an id that exceeds 64 characters, the parse succeeds since we can move the data into proto.  In this case, the issue will be reported to the ErrorReporter, but parsing will proceed, and can be considered successful since no data was lost.

In contrast, if a date has an invalid format, it will not be parsable to proto, the conversion will be lossy, and thus the parsing will be considered a failure (return false), even though the parser will continue to parse other fields.

To maintain the current APIs, top-level parse functions still return Status, and will return an InvalidArgument Status if the internal parse methods return failure.

Note that legacy behavior that makes use of FastFailErrorHandlers will be unaffected, since any report to those handlers will trigger a status failure, and thus the top-level parse functions will still return status failures.

PiperOrigin-RevId: 460544340