Add support for allowing references to base types of profiles.
Some references have urls of profiles in their list of allowed types. This change allows the dynamic validator to recognize those profiles and allow references that are base types of those profiles. e.g. if the profile is `us-core-patient`, both references to the `us-core-patient` profile and the `core-patient` resource should be allowed.

PiperOrigin-RevId: 496789108