Add support for user-defined callback functions
EvaluationResult can now represent an intermediate result when the expression has one or more callback functions. The caller would need to evaluate the function, given the input and call ResumeEvaluation to continue evaluating the expression.
CompiledExpression remains immutable and the results of each callback function are cached in the work_space stored in EvaluationResult.

A use case for this is implementing the resolve() method and letting the caller handle how it's implemented. resolve() would pass the references as an intermediate result, and the caller need to fetch the resources for those references and provide them as the result when calling ResumeEvaluation.

PiperOrigin-RevId: 588065917