Adds retagging logic to protogenerator to give fields whose datatype has been changed new tag numbers.
In other words, if Patient.foo used to be a Boolean in r4, and it is changed to a choice type of Boolean or CodeableConcept in r5, this is assigned a new tag number rather than reusing the old tag number.

This is designed ensure that there are no incompatible fields between versions, to ensure maximum flexibility for an ultimate versionless proto.

PiperOrigin-RevId: 600890279